requirements.txt
selenium
webdriver-manager
google-genai
# openai  # Uncomment if using OpenAI
# anthropic # Uncomment if using Anthropic
# python-dotenv
